sahil/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
google-generativeai==0.5.4
pdf2image==1.17.0
Pillow==10.3.0
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
supabase==2.4.0